Converting Walkthrough kit to Typescript
